refactor(suite): organize label files into labels/ dirs
Move wallet and address label selectors/hooks into src/labels/,
mirroring the existing structure in @suite/account.
