+++ suite/address/src/index.ts
@@ -1,5 +1,5 @@
 export * from './Address';
 export * from './copyAddressActions';
-export * from './selectAddressLabel';
-export * from './selectAddressLabelsForAccount';
-export * from './selectLabeledUnusedAddresses';
+export * from './labels/selectAddressLabel';
+export * from './labels/selectAddressLabelsForAccount';
+export * from './labels/selectLabeledUnusedAddresses';
